2021 go refactors & optimisations - Changed vec2.Vec2 to use int instead of int64 - Added some QoL methods to vec2.Vec for converting between 1d indexes - switched runner over to loov/hrtime so that measurements work on windows - Changed day 11 grids to []uint8 from map[vec2.Vec2]uint8 for a huge perf gain - Memoized neighbours in day 11 for even more perf

+++ go/aoc_2021/internal/day04/day04.go
@@ -107,7 +107,7 @@ func parseBoards(lines []string) []*board {
 			for col := 0; col < 5; col++ {
 				v, _ := strconv.ParseInt(strings.TrimSpace(l[col*3:col*3+2]), 10, 32)
 				unmarked += v
-				valMap[v] = vec2.Of(int64(col), int64(row))
+				valMap[v] = vec2.Of(col, row)
 			}
 			marked[row] = make([]bool, 5)
 		}
